+++ client/src/components/Shared/Modal/AddressModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AddressModal';
