+++ client/src/components/Shared/Modal/styles.ts
@@ -47,36 +47,3 @@ export const ModalButton = styled.button`
     opacity: 0.75;
   }
 `;
-
-export const PostcodeWrapper = styled.div`
-  margin-top: 1.2rem;
-  display: flex;
-  justify-content: space-between;
-  input {
-    flex: 2;
-    margin-top: 0 !important;
-    margin-left: 2rem;
-  }
-
-  button {
-    border-radius: 10px;
-    ${({ theme }) => theme.fontSize.m}
-  }
-`;
-
-export const DefaultAddrssCheckbox = styled(Checkbox)`
-  margin-top: 1.2rem;
-  color: ${({ theme }) => theme.color['text-color']};
-`;
-
-export const DuamPostWrapper = styled.div`
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  right: 0;
-  bottom: 0;
-  transform: translate(-50%, -50%);
-  width: 36rem;
-  height: 50rem;
-  background-color: rgba(1, 1, 1, 0.55);
-`;
